refactor: ♻️ extract CreateVesting logic into composables
CreateVesting.vue exceeded the 300-line max-lines rule. Move all
reactive form logic into a new useCreateVesting composable and split the
period/duration concern into useVestingDateRange. Export the TokenBalance
interface from useContractBalance so the composable's return type is
nameable across module boundaries.

Refs #2036
